# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Creatrid API
  description: |
    Creatrid is a verified digital identity platform for creators. Connect social
    accounts, build a Creator Score, manage content licensing, and share a public
    profile.

    ## Authentication
    Most endpoints require authentication via an HTTP-only cookie named `token`
    that contains a JWT issued after Google OAuth sign-in.

    Third-party verification endpoints use API key authentication via the
    `X-API-Key` header.
  version: 1.0.0
  contact:
    name: Creatrid
    url: https://creatrid.com
servers:
  - url: https://creatrid-api.bluewave-351c9003.westeurope.azurecontainerapps.io
    description: Production
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /api/auth/google/callback:
    get:
      operationId: googleCallback
      tags: [Auth]
      summary: Google OAuth callback
      description: |
        Handles the OAuth callback from Google. Validates state, exchanges the
        authorization code for tokens, finds or creates the user, issues a JWT
        cookie, and redirects to the frontend.
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
          description: Authorization code from Google
        - name: state
          in: query
          required: true
          schema:
            type: string
          description: CSRF state parameter
      responses:
        "307":
          description: Redirect to frontend (dashboard or onboarding)
...
  /api/analytics:
    get:
      operationId: getAnalyticsSummary
      tags: [Analytics]
      summary: Get analytics summary
      description: Returns profile view and click analytics for the authenticated user.
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Analytics summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsSummary"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/discover:
    get:
      operationId: discoverCreators
      tags: [Discovery]
      summary: Discover creators
      description: Returns a paginated list of creators with optional filters for score, platform, and search.
      parameters:
        - name: minScore
          in: query
          schema:
            type: integer
          description: Minimum Creator Score filter
        - name: platform
          in: query
          schema:
            type: string
            enum: [youtube, github, twitter, linkedin, instagram, dribbble, behance]
          description: Filter by connected platform
        - name: q
          in: query
          schema:
            type: string
          description: Search by name or username
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Creators list
          content:
            application/json:
              schema:
                type: object
                properties:
                  creators:
                    type: array
                    items:
                      $ref: "#/components/schemas/PublicUser"
                  total:
                    type: integer

  # ──────────────────────────────────────────────
  # Widget
  # ──────────────────────────────────────────────
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
      description: JWT token set as an HTTP-only cookie after Google OAuth sign-in
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for third-party verification endpoints (starts with crk_)
  parameters:
...
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
      description: Number of items to return (default 20, max 100)
    Offset:
      name: offset
      in: query
      schema:
        type: integer
        default: 0
        minimum: 0
      description: Number of items to skip
  responses:
    Unauthorized:
      description: Not authenticated
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: Not authenticated
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
...
    PublicUser:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
          nullable: true
        username:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        image:
          type: string
          nullable: true
        role:
          type: string
          enum: [CREATOR, BRAND, ADMIN]
        creatorScore:
          type: integer
          nullable: true
        creatorTier:
          type: string
        isVerified:
          type: boolean
        theme:
          type: string
        customLinks:
          type: array
          items:
            $ref: "#/components/schemas/CustomLink"
        createdAt:
          type: string
          format: date-time
    CustomLink:
      type: object
      properties:
        title:
          type: string
        url:
          type: string
          format: uri
...
    AnalyticsSummary:
      type: object
      description: Profile analytics summary with views and clicks broken down by day, referrer, browser, device, and location.
...